pyproject.toml
[project]
name = "document-extraction"
version = "0.1.0"
description = "POC d'extraction de données depuis des documents (factures, contrats)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Datalyo", email = "contact@example.com" }]
keywords = ["ocr", "document-extraction", "llm", "vertex-ai", "invoices", "contracts"]

dependencies = [
    # Core
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    # PDF/Images
    "pdf2image>=1.16.0",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    # OCR
    "pytesseract>=0.3.10",
    # LLM
    "google-cloud-aiplatform>=1.38.0",
    # Interface
    "streamlit>=1.28.0",
    # Data
    "pandas>=2.0.0",
    # Evaluation & Visualization
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "plotly>=5.18.0",
    # Notebook
    "jupyter>=1.0.0",
    "ipywidgets>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
]

[project.scripts]
extract = "document_extraction.pipeline:main"
evaluate = "evaluation.evaluate_performance:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/document_extraction"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "D", "UP"]
ignore = ["D100", "D104"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src/document_extraction"
